fix: moved JSONRPCErrorCode enums back to errors.ts to avoid use before bug
